add installed check

+++ softwares/apt/submodules/simplescreenrecorder.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 set -e
-sudo add-apt-repository ppa:maarten-baert/simplescreenrecorder -y &&
-    sudo apt-get update && sudo apt-get install simplescreenrecorder -y
+
+if [ ! -f /usr/bin/simplescreenrecorder ]; then
+    sudo add-apt-repository ppa:maarten-baert/simplescreenrecorder -y &&
+        sudo apt-get update && sudo apt-get install simplescreenrecorder -y
+fi
