+++ softwares/apt/submodules/system_monitor.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 set -e
-sudo add-apt-repository ppa:fossfreedom/indicator-sysmonitor -y &&
-    sudo apt-get update && sudo apt-get install indicator-sysmonitor -y
+
+if [ ! -f /usr/bin/indicator-sysmonitor ]; then
+    sudo add-apt-repository ppa:fossfreedom/indicator-sysmonitor -y &&
+        sudo apt-get update && sudo apt-get install indicator-sysmonitor -y
+fi
